+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/test.py
@@ -96,6 +96,7 @@ def test(
     root = get_root()
     testing_on_ci = running_on_ci()
     color = ctx.obj['color']
+    repo = ctx.obj['repo_name']
 
     # Implicitly track coverage
     if cov_missing:
@@ -286,7 +287,9 @@ def test(
                         abort('\nFailed!', code=result.code)
 
                     fix_coverage_report(check, 'coverage.xml')
-                    run_command(['codecov', '-X', 'gcov', '--root', root, '-F', check, '-f', 'coverage.xml'])
+
+                    if repo == 'core':
+                        run_command(['codecov', '-X', 'gcov', '--root', root, '-F', check, '-f', 'coverage.xml'])
                 else:
                     if not cov_keep:
                         remove_path('.coverage')