+++ ibm_mq/tests/test_ibm_mq_e2e.py
@@ -6,10 +6,11 @@ import pytest
 
 from datadog_checks.dev.utils import get_metadata_metrics
 
-from .common import MQ_VERSION, assert_all_metrics
+from .common import MQ_VERSION, assert_all_metrics, skip_windows_ci
+
+pytestmark = [skip_windows_ci, pytest.mark.e2e]
 
 
-@pytest.mark.e2e
 def test_e2e_check_all(dd_agent_check, instance_collect_all):
     aggregator = dd_agent_check(instance_collect_all, rate=True)
 
@@ -17,7 +18,6 @@ def test_e2e_check_all(dd_agent_check, instance_collect_all):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics())
 
 
-@pytest.mark.e2e
 @pytest.mark.skipif(
     MQ_VERSION < 9, reason='Only test for for version >=9, for v8 use a custom image with custom setup.'
 )